Split AppConfiguration data/factory concerns
