--- app-apple/Sources/CommonLibrary/Dependencies/AppConfiguration+Registry.swift
@@ -1,124 +1,0 @@
-// SPDX-FileCopyrightText: 2026 Davide De Rosa
-//
-// SPDX-License-Identifier: GPL-3.0
-
-import Partout
-
-extension ABI.AppConfiguration {
-    public func newRegistry(
-        deviceId: String,
-        cachesURL: URL,
-        configBlock: @escaping @Sendable () -> Set<ABI.ConfigFlag>
-    ) -> CodingRegistry {
-        let customHandlers: [ModuleHandler] = [
-            ProviderModule.moduleHandler
-        ]
-        let allImplementations: [ModuleImplementation] = [
-            OpenVPNImplementationBuilder(
-                distributionTarget: bundle.distributionTarget,
-                cachesURL: cachesURL,
-                configBlock: configBlock
-            ).build(),
-            WireGuardImplementationBuilder(
-                configBlock: configBlock
-            ).build()
-        ]
-        // Deprecated
-        var providerResolvers: [ProviderModuleResolver] = []
-        providerResolvers.append(OpenVPNProviderResolver())
-        providerResolvers.append(WireGuardProviderResolver(deviceId: deviceId))
-        let mappedResolvers = providerResolvers
-            .reduce(into: [:]) {
-                $0[$1.moduleType] = $1
-            }
-
-        let registry = Registry(
-            withKnown: true,
-            customHandlers: customHandlers,
-            allImplementations: allImplementations,
-            resolvedModuleBlock: {
-                do {
-                    return try Registry.resolvedModule($0, in: $1, with: mappedResolvers)
-                } catch {
-                    pspLog($1?.id, .core, .error, "Unable to resolve module: \(error)")
-                    throw error
-                }
-            }
-        )
-        registry.assertMissingImplementations()
-        return CodingRegistry(
-            registry: registry,
-            customModuleHandler: {
-                switch $0.innerType {
-                case .Provider:
-                    do {
-                        let data = try ABI.encode($0.json)
-                        return try ABI.decode(ProviderModule.self, from: data)
-                    } catch {
-                        pspLog(.profiles, .error, "Unable to decode ProviderModule: \(error)")
-                        return $0
-                    }
-                default:
-                    return $0
-                }
-            }
-        )
-    }
-}
-
-private extension Registry {
-    @Sendable
-    static func resolvedModule(
-        _ module: Module,
-        in profile: Profile?,
-        with resolvers: [ModuleType: ProviderModuleResolver]
-    ) throws -> Module {
-        do {
-            if let profile {
-                profile.assertSingleActiveProviderModule()
-                guard profile.isActiveModule(withId: module.id) else {
-                    return module
-                }
-            }
-            guard let providerModule = module as? ProviderModule else {
-                return module
-            }
-            guard let resolver = resolvers[providerModule.providerModuleType] else {
-                return module
-            }
-            return try resolver.resolved(from: providerModule)
-        } catch {
-            throw error as? PartoutError ?? PartoutProviderError.corruptModule(error)
-        }
-    }
-
-    func assertMissingImplementations() {
-        ModuleType.knownTypes.forEach { moduleType in
-            let builder = newModule(ofType: moduleType)
-            do {
-                // ModuleBuilder -> Module
-                let module = try builder.build()
-
-                // Module -> ModuleBuilder
-                guard let moduleBuilder = module.moduleBuilder() else {
-                    fatalError("\(moduleType): does not produce a ModuleBuilder")
-                }
-
-                // AppFeatureRequiring
-                guard builder is any AppFeatureRequiring else {
-                    fatalError("\(moduleType): #1 is not AppFeatureRequiring")
-                }
-                guard moduleBuilder is any AppFeatureRequiring else {
-                    fatalError("\(moduleType): #2 is not AppFeatureRequiring")
-                }
-            } catch {
-                switch (error as? PartoutError)?.code {
-                case .incompleteModule, .invalidField, .wireGuardEmptyPeers:
-                    return
-                default:
-                    fatalError("\(moduleType): empty module is not buildable: \(error)")
-                }
-            }
-        }
-    }
-}
